+++ .github/workflows/update.yml
@@ -2,7 +2,7 @@
 name: update
 on:
   push:
-    branches: main      
+    branches: main
 jobs:
   update:
     name: EAS Update
