+++ src/Components/Buttons/Haptics.tsx
