+++ src/Screens/Add.tsx
@@ -1,7 +1,7 @@
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import React from "react";
 import { StyleSheet, Text, TextInput, View } from "react-native";
-import GotoButton from "../Components/GotoButton";
+import GotoButton from "../Components/Buttons/GotoButton";
 import { RootStackParamList } from "../navigation/types";
 
 type Props = NativeStackScreenProps<RootStackParamList, "Add">;
