+++ src/Screens/Home.tsx
@@ -1,12 +1,10 @@
 import { NativeStackScreenProps } from "@react-navigation/native-stack";
 import React from "react";
 import { View } from "react-native";
-import GotoButton from "../Components/GotoButton";
-import HapticButton from "../Components/Haptics";
-import { RootStackParamList } from "../navigation/types";
+import GotoButton from "../Components/Buttons/GotoButton";
+import HapticButton from "../Components/Buttons/Haptics";
 import { useGameContext } from "../Context/GameContext";
-
-
+import { RootStackParamList } from "../navigation/types";
 
 type Props = NativeStackScreenProps<RootStackParamList, "Home">;
 
@@ -16,10 +14,15 @@ const Home = (navigator: Props) => {
   return (
     <View>
       <HapticButton title='GaG me!😘' to='PlayThis' navigator={navigator} hapticType='Heavy' />
-      {FavGames.length > 0 ? ( 
-       <HapticButton title='GaG a Fav😊' to='PlayThisFav' navigator={navigator} hapticType='Heavy' />) : null
-      }
-      
+      {FavGames.length > 0 ? (
+        <HapticButton
+          title='GaG a Fav😊'
+          to='PlayThisFav'
+          navigator={navigator}
+          hapticType='Heavy'
+        />
+      ) : null}
+
       <GotoButton title='List of your games' to='GameList' navigator={navigator} />
       <GotoButton title='List of your favorite games' to='FavGameList' navigator={navigator} />
     </View>
