+++ src/navigation/HomeStack.tsx
@@ -19,13 +19,16 @@ const RootStack = createNativeStackNavigator<RootStackParamList>();
 
 const HomeStackNavigator = () => {
   return (
-    <RootStack.Navigator  screenOptions={{ 
-      headerBackground: () => <View style={{ backgroundColor: '#876796', flex: 1, paddingLeft: 8 }} >
-      </View>,
-      headerRight: () => <HomeButton />,
-      headerLeft: () => <SettingsButton />,
-      headerTitle: () => <TitlePlacer />,      
-    }}>
+    <RootStack.Navigator
+      screenOptions={{
+        headerBackground: () => (
+          <View style={{ backgroundColor: "#876796", flex: 1, paddingLeft: 8 }}></View>
+        ),
+        headerRight: () => <HomeButton />,
+        headerLeft: () => <SettingsButton />,
+        headerTitle: () => <TitlePlacer />,
+      }}
+    >
       <RootStack.Screen name='Home' component={Home} />
       <RootStack.Screen name='Settings' component={Settings} />
       <RootStack.Screen
